Gracefully return not found if jest config not found
Since the resolver looks up the Jest config relative to the file
containing the import statement, it will not expect to find a
configuration when recursing through package dependencies. This is fine,
and just means that the resolver should return not found, allowing the
import plugin to fall back to a regular node resolver.

Unfortunately this does make tracking down problems where the
configuration actually is missing more difficult, but I don't see a way
to fix that without completely overhauling the way configuration lookup
is done.